refactor: unify platform code with #ifdef, add CMake + CI
- Merge linux_build/np_thread_linux.cpp into src/np_thread.cpp
  using #ifdef _MSC_VER for Windows FLS vs POSIX pthread_key
- Remove linux_build/ duplicated source files
- Add CMakeLists.txt for cross-platform build (MSVC + GCC/Clang)
- Add bench/bench.cpp as unified cross-platform benchmark
- Add .github/workflows/ci.yml (Windows + Linux build + benchmark)
- Fix local_new.h missing np_common.h include
