+++ .gitignore
@@ -2,6 +2,21 @@
 ## files generated by popular Visual Studio add-ons.
 
 
+# CMake build directory
+build/
+
+# linux_build artifacts
+linux_build/*.o
+linux_build/*.dSYM/
+linux_build/bench_*
+linux_build/np_bench*
+linux_build/np_alloc_bench
+linux_build/check_atomic
+linux_build/crash_test*
+linux_build/debug_test*
+linux_build/mini_test*
+linux_build/simple_test
+
 #내꺼
 nogit.*
 _[Bb]in/
