+++ bench/bench.cpp
@@ -1,3 +1,9 @@
+/****************************************************************************
+ *  bench.cpp — cross-platform np_alloc benchmark
+ *
+ *  Compares malloc vs np_alloc across single/multi-thread workloads.
+ *  Build via CMake: cmake --build build --target np_bench
+ ***/
 #include <chrono>
 #include <iostream>
 #include <random>
@@ -6,27 +12,38 @@
 #include <iomanip>
 #include <csignal>
 #include <cstdio>
-#include <unistd.h>
+#include <cstdlib>
 
-#include "../src/np_common.h"
-#include "../src/np_alloc.h"
+#include "np_alloc.h"
 
 using Clock = std::chrono::high_resolution_clock;
 
-void signal_handler(int sig) {
+#ifndef _MSC_VER
+#include <unistd.h>
+static void abort_handler(int sig) {
     fprintf(stderr, "Signal %d received!\n", sig);
     _exit(sig);
 }
+#else
+static void abort_handler(int sig) {
+    fprintf(stderr, "Signal %d received!\n", sig);
+    exit(sig);
+}
+#endif
+
+// ── Allocator API wrappers ──────────────────────────────────────────────
 
 struct malloc_API {
     static void* alloc(size_t sz) { return ::malloc(sz); }
-    static void free(void* ptr) { ::free(ptr); }
+    static void  free(void* ptr)  { ::free(ptr); }
 };
 
 struct np_alloc_API {
     static void* alloc(size_t sz) { return np_alloc(sz); }
-    static void free(void* ptr) { np_free(ptr); }
+    static void  free(void* ptr)  { np_free(ptr); }
 };
+
+// ── Single-thread benchmark ─────────────────────────────────────────────
 
 template <typename API>
 double run_single(int iterations, int max_alloc, int min_sz, int max_sz) {
@@ -57,6 +74,8 @@ double run_single(int iterations, int max_alloc, int min_sz, int max_sz) {
     auto end = Clock::now();
     return std::chrono::duration<double>(end - start).count();
 }
+
+// ── Multi-thread benchmark ──────────────────────────────────────────────
 
 template <typename API>
 double run_multi(int nthreads, int iterations, int max_alloc, int min_sz, int max_sz) {
@@ -93,29 +112,41 @@ double run_multi(int nthreads, int iterations, int max_alloc, int min_sz, int ma
     return std::chrono::duration<double>(end - start).count();
 }
 
+// ── Main ────────────────────────────────────────────────────────────────
+
 int main() {
-    signal(SIGSEGV, signal_handler);
-    signal(SIGBUS, signal_handler);
+    signal(SIGSEGV, abort_handler);
+#ifndef _MSC_VER
+    signal(SIGBUS, abort_handler);
+#endif
 
-    printf("=== np_alloc Benchmark (POSIX/ARM64) ===\n");
+    printf("=== np_alloc Benchmark ===\n");
 
-    // Warm up
+    // Warm up: ensure global pool is initialized
     void* w = np_alloc(64);
     np_free(w);
     printf("Warmup done.\n");
 
-    const int S_ITER = 500000;
+    const int S_ITER    = 500000;
     const int M_THREADS = 4;
-    const int M_ITER = 100000;
+    const int M_ITER    = 100000;
     const int MAX_ALLOC = 2000;
 
-    struct { const char* name; int thr; int iter; int lo; int hi; } tests[] = {
-        {"random_single  (100-7100)", 1, S_ITER, 100, 7100},
-        {"small_single   (50-300)",   1, S_ITER, 50, 300},
-        {"big_single     (5000-7500)",1, S_ITER, 5000, 7500},
-        {"random_multi   (100-7100)", M_THREADS, M_ITER, 100, 7100},
-        {"small_multi    (50-300)",   M_THREADS, M_ITER, 50, 300},
-        {"big_multi      (5000-7500)",M_THREADS, M_ITER, 5000, 7500},
+    struct TestCase {
+        const char* name;
+        int thr;
+        int iter;
+        int lo;
+        int hi;
+    };
+
+    TestCase tests[] = {
+        {"random_single  (100-7100)",  1,         S_ITER, 100,  7100},
+        {"small_single   (50-300)",    1,         S_ITER, 50,   300},
+        {"big_single     (5000-7500)", 1,         S_ITER, 5000, 7500},
+        {"random_multi   (100-7100)",  M_THREADS, M_ITER, 100,  7100},
+        {"small_multi    (50-300)",    M_THREADS, M_ITER, 50,   300},
+        {"big_multi      (5000-7500)", M_THREADS, M_ITER, 5000, 7500},
     };
 
     printf("\n%-32s %10s %10s %8s\n", "Test", "malloc", "np_alloc", "speedup");
@@ -124,17 +155,13 @@ int main() {
     for (auto& t : tests) {
         double mt, nt;
         if (t.thr == 1) {
-            printf("Running malloc %s...\n", t.name); fflush(stdout);
             mt = run_single<malloc_API>(t.iter, MAX_ALLOC, t.lo, t.hi);
-            printf("Running np_alloc %s...\n", t.name); fflush(stdout);
             nt = run_single<np_alloc_API>(t.iter, MAX_ALLOC, t.lo, t.hi);
         } else {
-            printf("Running malloc %s...\n", t.name); fflush(stdout);
             mt = run_multi<malloc_API>(t.thr, t.iter, MAX_ALLOC, t.lo, t.hi);
-            printf("Running np_alloc %s...\n", t.name); fflush(stdout);
             nt = run_multi<np_alloc_API>(t.thr, t.iter, MAX_ALLOC, t.lo, t.hi);
         }
-        printf("%-32s %8.4fs %8.4fs %6.2fx\n", t.name, mt, nt, mt/nt);
+        printf("%-32s %8.4fs %8.4fs %6.2fx\n", t.name, mt, nt, mt / nt);
     }
 
     printf("\n");
