--- linux_build/linux_stdafx.h
@@ -1,21 +1,0 @@
-// linux_stdafx.h — minimal stdafx.h for Linux build
-#pragma once
-
-#include <stdio.h>
-#include <stdint.h>
-#include <string.h>
-#include <assert.h>
-#include <stdlib.h>
-
-#include <string>
-#include <vector>
-#include <array>
-#include <set>
-
-#include <iostream>
-#include <sstream>
-
-#include <atomic>
-#include <functional>
-#include <utility>
-#include <tuple>
